Updated form handlers.

+++ src/HelioNetworks/FileManagerBundle/Form/DeleteFileFormHandler.php
@@ -7,17 +7,8 @@ use Symfony\Component\HttpFoundation\Request;
 
 use Symfony\Component\Form\Form;
 
-class DeleteFileFormHandler
+class DeleteFileFormHandler extends FormHandlers
 {
-    protected $form;
-    protected $request;
-
-    public function __construct(Form $form, Request $request)
-    {
-        $this->form = $form;
-        $this->request = $request;
-    }
-
     public function getFilename()
     {
         return $this->form->getData()->filename;
@@ -25,6 +16,11 @@ class DeleteFileFormHandler
 
     public function process(Filesystem $filesystem)
     {
-        $filesystem->delete($this->getFilename());
+        if($this->isValid()) {
+            return $filesystem->delete($this->getFilename());
+        }
+        else {
+            return false;
+        }
     }
 }
